blt-lettuce/pyproject.toml
[tool.poetry]
name = "lettuce-blt"
version = "0.1.0"
description = ""
authors = ["OWASP BLT Team", "Sarthak5598 <sarthak5598sharma@gmail.com>"]
readme = "README.md"
packages = [{ include = "lettuce", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
gitpython = "^3.1.43"
slack-machine = "^0.37.0"
slackeventsapi = "^3.0.1"
slack-sdk = "^3.27.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
pytest-mock = "^3.14.0"
ruff = "^0.4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
known_first_party = ["lettuce", "tests"]
line_length = 99
multi_line_output = 3
profile = "black"

[tool.ruff]
line-length = 99
target-version = "py311"

[tool.ruff.lint]
select = ["E4", "E5", "E7", "E9", "F", "N", "W"]

[tool.ruff.lint.flake8-errmsg]
max-string-length = 99
